+++ CHANGELOG.md
@@ -2,7 +2,20 @@
 
 All notables changes between versions are documented in this file.
 
-## v0.4.1 - 2025-08-??
+## v0.4.2 - 2025-09-06
+
+Added [One ROM Lab](rust/lab/README.md) support, which allows a One ROM to be used as a ROM reader.
+
+### Changes
+
+- Added fw-parser for One ROM Lab
+- Modify sdrr-info/parser to support Airfrog custom firmware changes
+
+### Fixes
+
+- Probably a few here and there.
+
+## v0.4.1 - 2025-08-28
 
 ### Changes